some refactoring for clarity
* rename `Kolibri` to `KolibriGlobalVar`
* move `kolibri_module` definition to `core`
* flattened directory structure in `core/assets/src`
* renamed `kolibri_core_app` to just `core_app`
* removed extra `facade` layer on the `core_app`
