+++ .eslintignore
@@ -1,5 +1,5 @@
 # TODO - refactor these files such that they pass eslint, and remove from here
 
-kolibri/core/assets/src/mediator/asset_loader.js
-kolibri/core/assets/src/mediator/mediator.js
-kolibri/core/assets/src/facade/facade.js
+kolibri/core/assets/src/asset_loader.js
+kolibri/core/assets/src/core_app_mediator.js
+kolibri/core/assets/src/core_app.js
