+++ kolibri/core/assets/src/asset_loader.js
