+++ kolibri/core/assets/src/core_app_mediator.js
