--- kolibri/core/assets/src/kolibri_core_app.js
@@ -1,12 +1,0 @@
-
-const Kolibri = require('./facade/facade');
-
-const logging = require('loglevel');
-
-const kolibri = new Kolibri();
-
-logging.setDefaultLevel(2);
-
-logging.info('Kolibri Initialized');
-
-module.exports = kolibri;
